Move `zero_point_flux` equivalency implementation
The move should be invisible to users because they should be importing
the equivalency from the `astropy.units` namespace both before and after
this move.
