+++ astropy/units/photometric.py
@@ -18,10 +18,10 @@ import numpy as np
 from astropy.constants.si import L_bol0
 
 from . import astrophys, cgs, si
-from .core import Unit, UnitBase, def_unit
+from .core import UnitBase, def_unit
 from .utils import generate_unit_summary
 
-__all__ = ["zero_point_flux"]  #  Units are added at the end
+__all__ = []  #  Units are added at the end
 
 _ns = globals()
 
@@ -69,20 +69,6 @@ def_unit(
         "zero_point_flux equivalency should be used."
     ),
 )
-
-
-def zero_point_flux(flux0):
-    """
-    An equivalency for converting linear flux units ("maggys") defined relative
-    to a standard source into a standardized system.
-
-    Parameters
-    ----------
-    flux0 : `~astropy.units.Quantity`
-        The flux of a magnitude-0 object in the "maggy" system.
-    """
-    flux_unit0 = Unit(flux0)
-    return [(maggy, flux_unit0)]
 
 
 ###########################################################################
